requirements.txt
discord.py[voice]>=2.3.2
yt-dlp>=2024.1.0
spotipy>=2.23.0
PyNaCl>=1.5.0
python-dotenv>=1.0.0
